+++ help/source/conf.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# TelemacTools documentation build configuration file, created by
+# MeshTools documentation build configuration file, created by
 # sphinx-quickstart on Sun Feb 12 17:11:03 2012.
 #
 # This file is execfile()d with the current directory set to its containing dir.
@@ -41,7 +41,7 @@ source_suffix = ".rst"
 master_doc = "index"
 
 # General information about the project.
-project = "TelemacTools"
+project = "MeshTools"
 copyright = "2013, APn/Artelia"
 
 # The version info for the project you're documenting, acts as replacement for
@@ -179,7 +179,7 @@ htmlhelp_basename = "TemplateClassdoc"
 # Grouping the document tree into LaTeX files. List of tuples
 # (source start file, target name, title, author, documentclass [howto/manual]).
 latex_documents = [
-    ("index", "TelemacTools.tex", "TelemacTools Documentation", "APn/Artelia", "manual"),
+    ("index", "MeshTools.tex", "MeshTools Documentation", "APn/Artelia", "manual"),
 ]
 
 # The name of an image file (relative to this directory) to place at the top of
@@ -210,4 +210,4 @@ latex_documents = [
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [("index", "TemplateClass", "TelemacTools Documentation", ["APn/Artelia"], 1)]
+man_pages = [("index", "TemplateClass", "MeshTools Documentation", ["APn/Artelia"], 1)]
