+++ help/source/index.rst
@@ -1,9 +1,9 @@
-.. TelemacTools documentation master file, created by
+.. MeshTools documentation master file, created by
    sphinx-quickstart on Sun Feb 12 17:11:03 2012.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to TelemacTools's documentation!
+Welcome to MeshTools's documentation!
 ============================================
 
 Contents:
