+++ mesh_tools/README.md
@@ -1,10 +1,11 @@
-# Telemac Tools 
-Tools for Telemac using Mesh format
+# Mesh Tools 
+Tools for management of data using Mesh format for Telemac or Uhaina
 More informations on Telemac could be found on [Open Telemac-Mascaret website]("http://www.opentelemac.org/")
+More informations on Uhaina could be found (soon)
 ## Installation</h3>
-Use the QGIS Plugins menu to install the Telemac Tools Plugin
+Use the QGIS Plugins menu to install the Mesh Tools Plugin
 
 ##License
-Telemac_Tools plugin is free software.
+Mesh_Tools plugin is free software.
 You can redistribute it anf/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation.
