+++ mesh_tools/__init__.py
@@ -1,15 +1,15 @@
 # -*- coding: utf-8 -*-
 """
 /***************************************************************************
- TelemacToolsDockWidget
+ MeshTools
                                  A QGIS plugin
- Tools for Telemac
+ Tools for management of Data on mesh (Telemac, Uhaina)
  Generated by Plugin Builder: http://g-sherman.github.io/Qgis-Plugin-Builder/
                              -------------------
         begin                : 2021-03-24
-        git sha              : $Format:%H$
         copyright            : (C) 2021 by APn/Artelia
         email                : a@a
+        git sha              : $Format:%H$
  ***************************************************************************/
 
 /***************************************************************************
@@ -20,30 +20,18 @@
  *   (at your option) any later version.                                   *
  *                                                                         *
  ***************************************************************************/
+ This script initializes the plugin, making it known to QGIS.
 """
 
-import os
 
-from qgis.PyQt import QtGui, QtWidgets, uic
-from qgis.PyQt.QtCore import pyqtSignal
+# noinspection PyPep8Naming
+def classFactory(iface):  # pylint: disable=invalid-name
+    """Load MeshTools class from file MeshTools.
 
-FORM_CLASS, _ = uic.loadUiType(os.path.join(os.path.dirname(__file__), "telemac_tools_dockwidget_base.ui"))
+    :param iface: A QGIS interface instance.
+    :type iface: QgsInterface
+    """
+    #
+    from .mesh_tools import MeshTools
 
-
-class TelemacToolsDockWidget(QtWidgets.QDockWidget, FORM_CLASS):
-
-    closingPlugin = pyqtSignal()
-
-    def __init__(self, parent=None):
-        """Constructor."""
-        super(TelemacToolsDockWidget, self).__init__(parent)
-        # Set up the user interface from Designer.
-        # After setupUI you can access any designer object by doing
-        # self.<objectname>, and you can use autoconnect slots - see
-        # http://doc.qt.io/qt-5/designer-using-a-ui-file.html
-        # #widgets-and-dialogs-with-auto-connect
-        self.setupUi(self)
-
-    def closeEvent(self, event):
-        self.closingPlugin.emit()
-        event.accept()
+    return MeshTools(iface)
