+++ mesh_tools/libs/__init__.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 """
 /***************************************************************************
- TelemacTools
+ MeshTools
                                  A QGIS plugin
- Tools for Telemac
+ Tools for  Tools for management of Data on mesh (Telemac, Uhaina)
  Generated by Plugin Builder: http://g-sherman.github.io/Qgis-Plugin-Builder/
                              -------------------
         begin                : 2021-03-24
@@ -26,12 +26,12 @@
 
 # noinspection PyPep8Naming
 def classFactory(iface):  # pylint: disable=invalid-name
-    """Load TelemacTools class from file TelemacTools.
+    """Load MeshTools class from file MeshTools.
 
     :param iface: A QGIS interface instance.
     :type iface: QgsInterface
     """
     #
-    from .telemac_tools import TelemacTools
+    from .mesh_tools import MeshTools
 
-    return TelemacTools(iface)
+    return MeshTools(iface)
