+++ mesh_tools/mesh_tools.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 """
 /***************************************************************************
- TelemacTools
+ MeshTools
                                  A QGIS plugin
- Tools for Telemac
+ Tools for management of Data on mesh (Telemac, Uhaina)
  Generated by Plugin Builder: http://g-sherman.github.io/Qgis-Plugin-Builder/
                               -------------------
         begin                : 2021-03-24
@@ -36,7 +36,7 @@ from .libs.mesh_quality import MeshQuality
 # from .resources import *
 
 
-class TelemacTools:
+class MeshTools:
     """QGIS Plugin Implementation."""
 
     def __init__(self, iface):
@@ -56,7 +56,7 @@ class TelemacTools:
 
         # initialize locale
         locale = QSettings().value("locale/userLocale")[0:2]
-        locale_path = os.path.join(self.plugin_dir, "i18n", "TelemacTools_{}.qm".format(locale))
+        locale_path = os.path.join(self.plugin_dir, "i18n", "MeshTools_{}.qm".format(locale))
 
         if os.path.exists(locale_path):
             self.translator = QTranslator()
@@ -65,13 +65,13 @@ class TelemacTools:
 
         # Declare instance attributes
         self.actions = []
-        self.menu = self.tr("&Telemac Tools")
+        self.menu = self.tr("&Mesh Tools")
         # TODO: We are going to let the user set this up in a future iteration
         self.toolbar = QToolBar()
-        # self.toolbar = self.iface.addToolBar(u'TelemacTools')
-        # self.toolbar.setObjectName(u'TelemacTools')
+        # self.toolbar = self.iface.addToolBar(u'MeshTools')
+        # self.toolbar.setObjectName(u'MeshTools')
 
-        # print "** INITIALIZING TelemacTools"
+        # print "** INITIALIZING MeshTools"
 
         self.pluginIsActive = False
         self.dockwidget = None
@@ -89,7 +89,7 @@ class TelemacTools:
         :rtype: QString
         """
         # noinspection PyTypeChecker,PyArgumentList,PyCallByClass
-        return QCoreApplication.translate("TelemacTools", message)
+        return QCoreApplication.translate("MeshTools", message)
 
     def add_action(
         self,
@@ -167,7 +167,7 @@ class TelemacTools:
         """Create the menu entries and toolbar icons inside the QGIS GUI."""
         # self.postAct.triggered.connect(lambda: self.run("post"))
         icon_path = os.path.join(self.plugin_dir, "icon.png")
-        # icon_path = ':/plugins/telemac_tools/icon.png'
+        # icon_path = ':/plugins/mesh_tools/icon.png'
         self.add_action(
             icon_path, text=self.tr("Culvert Manager"), callback=lambda: self.run(1), parent=self.iface.mainWindow()
         )
@@ -179,7 +179,7 @@ class TelemacTools:
                 parent=self.iface.mainWindow(),
             )
         self.add_action(
-            icon_path, text=self.tr("Telemac Tool 2"), callback=lambda: self.run(3), parent=self.iface.mainWindow()
+            icon_path, text=self.tr("Mesh Tool 2"), callback=lambda: self.run(3), parent=self.iface.mainWindow()
         )
 
     # --------------------------------------------------------------------------
@@ -203,10 +203,10 @@ class TelemacTools:
     def unload(self):
         """Removes the plugin menu item and icon from QGIS GUI."""
 
-        # print "** UNLOAD TelemacTools"
+        # print "** UNLOAD MeshTools"
 
         for action in self.actions:
-            self.iface.removePluginMenu(self.tr("&Telemac Tools"), action)
+            self.iface.removePluginMenu(self.tr("&Mesh Tools"), action)
             self.iface.removeToolBarIcon(action)
         # remove the toolbar
         del self.toolbar
@@ -226,7 +226,7 @@ class TelemacTools:
             #    removed on close (see self.onClosePlugin method)
             if self.dockwidget == None:
                 # Create the dockwidget (after translation) and keep reference
-                self.dockwidget = TelemacToolDockWidget()
+                self.dockwidget = MeshToolDockWidget()
 
                 # connect to provide cleanup on closing of dockwidget
                 self.dockwidget.closingPlugin.connect(self.onClosePlugin)
@@ -240,26 +240,26 @@ class TelemacTools:
             print(self.dockwidget.widget().close())
 
         if tool == 1:
-            self.dockwidget.setWindowTitle("Telemac - Culvert Manager")
+            self.dockwidget.setWindowTitle("Mesh - Culvert Manager")
             self.dockwidget.setWidget(CulvertManager())
         elif tool == 2:
-            self.dockwidget.setWindowTitle("Telemac - Mesh Quality Analysis")
+            self.dockwidget.setWindowTitle("Mesh - Mesh Quality Analysis")
             self.dockwidget.setWidget(MeshQuality())
         elif tool == 3:
-            self.dockwidget.setWindowTitle("Telemac - Tool {}".format(tool))
+            self.dockwidget.setWindowTitle("Mesh - Tool {}".format(tool))
             self.dockwidget.setWidget(QWidget())
         else:
-            self.dockwidget.setWindowTitle("Telemac - Erreur")
+            self.dockwidget.setWindowTitle("Mesh - Erreur")
             self.dockwidget.setWidget(QWidget())
 
         self.dockwidget.widget().show()
 
 
-class TelemacToolDockWidget(QDockWidget):
+class MeshToolDockWidget(QDockWidget):
     closingPlugin = pyqtSignal()
 
     def __init__(self, parent=None):
-        super(TelemacToolDockWidget, self).__init__(parent)
+        super(MeshToolDockWidget, self).__init__(parent)
 
     def closeEvent(self, event):
         self.closingPlugin.emit()
