+++ mesh_tools/mesh_tools_dockwidget_base.ui
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>TelemacToolsDockWidgetBase</class>
- <widget class="QDockWidget" name="TelemacToolsDockWidgetBase">
+ <class>MeshToolsDockWidgetBase</class>
+ <widget class="QDockWidget" name="MeshToolsDockWidgetBase">
   <property name="geometry">
    <rect>
     <x>0</x>
@@ -11,7 +11,7 @@
    </rect>
   </property>
   <property name="windowTitle">
-   <string>Telemac Tools</string>
+   <string>Mesh Tools</string>
   </property>
   <widget class="QWidget" name="dockWidgetContents">
    <layout class="QGridLayout" name="gridLayout">
