+++ mesh_tools/metadata.txt
@@ -3,14 +3,14 @@
 # This file should be included when you package your plugin.# Mandatory items:
 
 [general]
-name=Telemac Tools
+name=Mesh Tools
 qgisMinimumVersion=3.16
-description=Tools for Telemac(http://www.opentelemac.org/).
+description=Tools for management of Data on mesh (Telemac, Uhaina).
 version=0.3
 author=Artelia
 email=a@a
 
-about=Tools for Telemac result using Mesh format :
+about=Tools for for management of Data using Mesh format :
       - Culvert Manager developed by Aurélien Perrin(Artelia)
       - Mesh Quality Analysis developed by Nicolas Godet (ISL)
 
@@ -23,7 +23,7 @@ repository=https://github.com/Artelia/telemac_tools
 hasProcessingProvider=no
 # Uncomment the following line and add your changelog:
 changelog=
-   0.3   : Restructuration of github floders 
+   0.3   : Restructuration of github folders 
    0.2   : add Mesh quality analysis (Qgis >= 3.22.0)
