+++ pb_tool.cfg
@@ -1,5 +1,5 @@
 #/***************************************************************************
-# TelemacTools
+# MeshTools
 #
 # Configuration file for plugin builder tool (pb_tool)
 # Generated by Plugin Builder: http://g-sherman.github.io/Qgis-Plugin-Builder/
@@ -39,7 +39,7 @@
 [plugin]
 # Name of the plugin. This is the name of the directory that will
 # be created in .qgis2/python/plugins
-name: telemac_tools
+name: mesh_tools
 
 # Full path to where you want your plugin directory copied. If empty,
 # the QGIS default path will be used. Don't include the plugin name in
@@ -48,10 +48,10 @@ plugin_path:
 
 [files]
 # Python  files that should be deployed with the plugin
-python_files: __init__.py telemac_tools.py telemac_tools_dockwidget.py
+python_files: __init__.py mesh_tools.py mesh_tools_dockwidget.py
 
 # The main dialog file that is loaded (not compiled)
-main_dialog: telemac_tools_dockwidget_base.ui
+main_dialog: mesh_tools_dockwidget_base.ui
 
 # Other ui files for dialogs you create (these will be compiled)
 compiled_ui_files:
